Move graphql server to fix cloud tests
