+++ packages/gatsby-cli/src/recipes.ts
@@ -1,6 +1,6 @@
 import { trackCli } from "gatsby-telemetry"
 // NOTE(@mxstbr): I explicitly import from /index because the file used to live at graphql.js, which means developers with old builds on their local machines will have that old version imported instead of the new one with changes
-import startGraphQLServer from "gatsby-recipes/dist/graphql/index.js"
+import startGraphQLServer from "gatsby-recipes/dist/graphql-server"
 
 export async function recipesHandler(
   projectRoot: string,
