+++ packages/gatsby/src/commands/develop.ts
@@ -9,8 +9,7 @@ import socket from "socket.io"
 import fs from "fs-extra"
 import { isCI, slash } from "gatsby-core-utils"
 import { createServiceLock } from "gatsby-core-utils/dist/service-lock"
-// NOTE(@mxstbr): I explicitly import from /index because the file used to live at graphql.js, which means developers with old builds on their local machines will have that old version imported instead of the new one with changes
-import startGraphQLServer from "gatsby-recipes/dist/graphql/index.js"
+import startGraphQLServer from "gatsby-recipes/dist/graphql-server"
 import { startDevelopProxy } from "../utils/develop-proxy"
 import { IProgram } from "./types"
